Additional refactoring. Unfortunately breaks root task times.
With these changes the root task times are constantly being reset due to
having no proper "root" task time established. Instead they are set when
the thread is acquired, but lost just as soon as the threads are
released. Needs to be fixed with a different solution, probably
something like was in the previous iteration.